pyproject.toml
[project]
name = "internet_concierge"
version = "0.0.1"
description = "A helpful agent that uses the Internet for you"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "langsmith>=0.2.10",
    "playwright>=1.49.1",
    "python-dotenv>=1.0.1",
]
